% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSiMFD.R
\name{pred_dcompSiMVFD}
\alias{pred_dcompSiMVFD}
\title{Give the projection on the reduced dimension obtained from RSIMVFD decomposition}
\usage{
pred_dcompSiMVFD(fitOBJ, testFDATA, test_indxJ, test_spat_indx)
}
\arguments{
\item{fitOBJ}{an object obtained by the dredSIMVFD}

\item{testFDATA}{is a list of functional data from the test for which we want 
to obtain the projection onto the reduced dimension}

\item{test_indxJ}{is similar to indxJ, however, it corresponds to the test data provided}

\item{test_spat_index}{is a L by 2 matrix contains spatial locations for the test data
contains information on mi}
}
\value{
a list with following objects:
\itemize{
 \item n : number of subjects in the test data
 \item test_SRF : extracted spatial random fields
 \item test_factor_subtract : predicted subject-specific mean
 \item testRD : derived features from extracted SRF
}
}
\description{
Give the projection on the reduced dimension obtained from RSIMVFD decomposition
}
