% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSiMFD.R
\name{smth_covar}
\alias{smth_covar}
\title{Smooth covariance and cross-covariance for a zero-mean multivariate functional data}
\usage{
smth_covar(funDATA, funARG, pve, nKnot, basisT, tps_m)
}
\arguments{
\item{funDATA}{a list with elements as univariate observed functional data}

\item{funARG}{arguments of the observed function}

\item{pve}{a scalar represents the threshold for percentage of variation 
explained in the univariate FPCA for smoothing auto-covariance}

\item{nKnot}{a scalar representing the number of knots to be used in fpca.face and tensor product smoothing for cross-covariance}

\item{basisT}{is a vector of two elements contains the types of basis be used in tensor product smoothing}

\item{tps_m}{order of the penaly for tensor-product smoothing for cross-covariance}
}
\description{
Smooth covariance and cross-covariance for a zero-mean multivariate functional data
}
